requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
jinja2==3.1.4
python-multipart==0.0.9
sqlalchemy==2.0.34
psycopg2-binary==2.9.9
alembic==1.13.2
passlib[bcrypt]==1.7.4
pyjwt==2.9.0
pydantic==2.8.2
pydantic-settings==2.4.0

scikit-learn==1.5.1

pytest==8.3.2
httpx==0.27.2